Define LIV_TUT_NO_BEGIN_SHR_CLEAN and remove all the files inside the shared directory when this value is '1'

+++ provision-shell/tutorials/begin/begin.sh
@@ -51,5 +51,7 @@ sudo dpkg-reconfigure -f noninteractive tzdata > /dev/null 2>&1
 # versions of packages and their dependencies
 sudo apt-get -qq update
 
-# Remove all files from the shared directory
-sudo rm -rf "$LIV_DIR_SHR"/*
+# Remove all the files inside the shared directory
+if [ "$LIV_TUT_NO_BEGIN_SHR_CLEAN" -eq '1' ]; then
+    sudo rm -rf "$LIV_DIR_SHR"/*
+fi
